refactor(desktop): consolidate scrobbler into main/scrobble
